% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/borestapi-package.R
\docType{package}
\name{borestapi-package}
\alias{borestapi}
\alias{borestapi-package}
\title{borestapi: Interface to SAP Business Objects REST API}
\description{
An R interface to the SAP Business Objects REST API. Use it to: Retrieve data from Business Object document report tabs. Upload Excel spreadsheets to Business Objects Retrieve data from a Busines Objects document data source Set document filters and refresh document data
}
\author{
\strong{Maintainer}: John Capehart \email{R@jcapehart.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
