% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{check_bo_connection_state}
\alias{check_bo_connection_state}
\title{Check to see if an SAP access token is still valid}
\usage{
check_bo_connection_state(conn, server)
}
\arguments{
\item{conn}{connection reference}

\item{server}{Server name and port to connect to}
}
\value{
TRUE if connection is valid, FALSE otherwise
}
\description{
Check to see if an SAP access token is still valid
}
