% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{close_bo_connection}
\alias{close_bo_connection}
\title{Close connectionto BO server}
\usage{
close_bo_connection(conn)
}
\arguments{
\item{conn}{Connection reference to close}
}
\value{
Connection reference
}
\description{
Close connectionto BO server
}
