% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document.R
\name{close_bo_document}
\alias{close_bo_document}
\title{Close BO Webi document}
\usage{
close_bo_document(conn, document, save = FALSE)
}
\arguments{
\item{conn}{Connection reference}

\item{document}{Document as numeric id or tibble of properties}

\item{save}{If true save the document before closing}
}
\value{
Response content
}
\description{
Close BO Webi document
}
