% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document.R
\name{copyBODocument}
\alias{copyBODocument}
\title{Copy a Webi document}
\usage{
copyBODocument(conn, document, folder_id, newName)
}
\arguments{
\item{conn}{Connection reference}

\item{folder_id}{Numeric id of parent folder for copy}

\item{newName}{Name of new document}

\item{document_id}{Numeric id of Webi document}
}
\value{
Response content
}
\description{
Copy a Webi document
}
