% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document.R
\name{create_bo_document}
\alias{create_bo_document}
\title{Create a new webi document}
\usage{
create_bo_document(conn, name, parent_folder)
}
\arguments{
\item{conn}{Connection reference}

\item{name}{name of document}

\item{parent_folder}{Numeric id of parent folder for document}
}
\value{
Response content
}
\description{
Create a new webi document
}
