% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spreadsheet.R
\name{DELETE_bo_spreadsheet}
\alias{DELETE_bo_spreadsheet}
\title{Delete a spreadsheet from BO}
\usage{
DELETE_bo_spreadsheet(conn, filename, folder_id)
}
\arguments{
\item{conn}{Connection reference}

\item{filename}{Name of the spreadsheet}

\item{folder_id}{Numeric id of the folder}
}
\value{
Response content
}
\description{
Delete a spreadsheet from BO
}
