% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/password.R
\name{delete_keyring_secret}
\alias{delete_keyring_secret}
\title{Delete a secret from the keyring}
\usage{
delete_keyring_secret(name)
}
\arguments{
\item{name}{Secret name}
}
\description{
Delete a secret from the keyring
}
