% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controls.R
\name{get_bo_control_details}
\alias{get_bo_control_details}
\title{Get input controls for document or report}
\usage{
get_bo_control_details(
  conn,
  document,
  report = NULL,
  control = "",
  all_info = FALSE
)
}
\arguments{
\item{conn}{Connection reference}

\item{document}{Document as numeric id or tibble of properties}

\item{report}{Report name (omit for document level controls)}

\item{control}{Control name (omit to return all controls)}

\item{all_info}{Return all control details}
}
\value{
Controls as tibble
}
\description{
Get input controls for document or report
}
