% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controls.R
\name{get_bo_control_selection}
\alias{get_bo_control_selection}
\title{Get the selection state of a document control}
\usage{
get_bo_control_selection(conn, document, control_name)
}
\arguments{
\item{conn}{Connection reference}

\item{document}{Document as numeric id or tibble of properties}

\item{control_name}{Name of control}
}
\value{
'@all' or tibble of selected levels
}
\description{
Get the selection state of a document control
}
