% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controls.R
\name{get_bo_control_selection_set}
\alias{get_bo_control_selection_set}
\title{Title}
\usage{
get_bo_control_selection_set(conn, document, report = "", control = "")
}
\arguments{
\item{conn}{Connection reference}

\item{document}{Document as numeric id or tibble of properties}

\item{report}{Identifier of report for report level controls, blank otherwise}

\item{control}{Identifier of control}
}
\value{
Control selection set as tibble
}
\description{
Title
}
