% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document.R
\name{get_bo_data_provider_data}
\alias{get_bo_data_provider_data}
\title{Get data directly from a document data provider}
\usage{
get_bo_data_provider_data(conn, document, data_provider, ...)
}
\arguments{
\item{conn}{Connection reference}

\item{document}{Document as numeric id or tibble of properties}

\item{data_provider}{Identifier of the provider}

\item{...}{Parameters for read_delim}
}
\value{
Tibble of data
}
\description{
See also get_bo_report_data
}
