% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document.R
\name{get_bo_data_provider_details}
\alias{get_bo_data_provider_details}
\title{Title}
\usage{
get_bo_data_provider_details(conn, document, data_provider = "")
}
\arguments{
\item{conn}{Connection Reference}

\item{document}{Document as numeric id or tibble of properties}

\item{data_provider}{Name of the data provider (optional)}
}
\value{
Provider detals as tibble
}
\description{
Title
}
