% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controls.R
\name{get_bo_data_provider_specification}
\alias{get_bo_data_provider_specification}
\title{Get the date range for a data provider specification}
\usage{
get_bo_data_provider_specification(conn, document, dataprovider)
}
\arguments{
\item{conn}{Connection reference}

\item{document}{Document as numeric id or tibble of properties}

\item{dataprovider}{Name of data provider from document}
}
\value{
XML string of data provider specification
}
\description{
Get the date range for a data provider specification
}
