% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document.R
\name{get_bo_document}
\alias{get_bo_document}
\title{Title}
\usage{
get_bo_document(conn, document)
}
\arguments{
\item{conn}{Connection Reference}

\item{document}{Document as numeric id or tibble of properties}
}
\value{
Document properties as tibble
}
\description{
Title
}
