% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document.R
\name{get_bo_document_control_selection_set}
\alias{get_bo_document_control_selection_set}
\title{Title}
\usage{
get_bo_document_control_selection_set(conn, document, controlName = NULL)
}
\arguments{
\item{conn}{Connection reference}

\item{document}{Document as numeric id or tibble of properties}

\item{controlName}{}
}
\description{
Title
}
