% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document.R
\name{get_bo_document_data}
\alias{get_bo_document_data}
\title{Get data directly from a document data provider
See also get_bo_document_report}
\usage{
get_bo_document_data(conn, document, provider_id, ...)
}
\arguments{
\item{conn}{Connection reference}

\item{document}{Document as numeric id or tibble of properties}

\item{provider_id}{Numeric id of the provider}

\item{...}{Parameters for read_delim}
}
\value{
Tible of data
}
\description{
Get data directly from a document data provider
See also get_bo_document_report
}
