% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document.R
\name{get_bo_document_details}
\alias{get_bo_document_details}
\title{Get the meta data for a Webi document}
\usage{
get_bo_document_details(conn, document)
}
\arguments{
\item{conn}{Connection Reference}

\item{document}{Document as numeric id or tibble of properties}
}
\value{
Document properties as tibble
}
\description{
Get the meta data for a Webi document
}
