% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spreadsheet.R
\name{get_bo_folder_children}
\alias{get_bo_folder_children}
\title{Get children of a folder}
\usage{
get_bo_folder_children(conn, parent_folder)
}
\arguments{
\item{conn}{Connection reference}

\item{parent_folder}{id of parent folder}
}
\value{
tibble with chilren properties
}
\description{
Get children of a folder
}
