% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-utils.R
\name{get_bo_item}
\alias{get_bo_item}
\title{Retrieve BO item proerties}
\usage{
get_bo_item(
  conn,
  name = NULL,
  parent_folder = NULL,
  kind = NULL,
  owner = NULL,
  id = NULL
)
}
\arguments{
\item{conn}{Connection reference}

\item{name}{Name of the item (optional). You can use SQL wilcard \% in the name.}

\item{parent_folder}{Numeric id of the parent folder (optional)}

\item{kind}{Kind of item (optional)}

\item{owner}{Owner of item (optional)}

\item{id}{Id of item (optional)}
}
\value{
Tibble of item properties
}
\description{
Get the properties of items as tibble.
}
\details{
After opening a connection with open_bo_connection you can retrieve the properties of a items with this function.
The properties returned contain the item id in the SI_ID column. This matches the ID in the properties page in the BO web GUI.
If you have the id of the parent folder that will narrow the search in case there are duplicate item names. If more
than one item is returned by the query the tibble will contain multiple rows
}
\examples{
# get the id of an item. Normally you would call open_bo_connection()
conn= open_bo_connection(); get_bo_item(conn,name='test'),folder=3944223)$SI_ID
# get all the Excel documents in a folder
get_bo_item(open_bo_connection(),parent_folder = 3944223, kind = 'Excel'))
}
