% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-utils.R
\name{get_bo_item_from_id}
\alias{get_bo_item_from_id}
\title{Title}
\usage{
get_bo_item_from_id(conn, id)
}
\arguments{
\item{conn}{Connection Reference}

\item{id}{Numeric id of item}
}
\value{
Item as a tibble of properties
}
\description{
Title
}
