% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-utils.R
\name{get_bo_item_from_name}
\alias{get_bo_item_from_name}
\title{Retreive BO item by name}
\usage{
get_bo_item_from_name(conn, name, folder_id, kind = NULL, owner = NULL)
}
\arguments{
\item{conn}{Connection reference}

\item{name}{Name of the item}

\item{folder_id}{Numeric id of the parent folder}

\item{kind}{kind of item (optional)}

\item{owner}{owner of item (optional)}
}
\value{
Tibble of item properties
}
\description{
Retreive BO item by name
}
