% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-utils.R
\name{get_bo_item_from_pattern}
\alias{get_bo_item_from_pattern}
\title{Retrieve BO object from an matching pattern}
\usage{
get_bo_item_from_pattern(
  conn,
  pattern,
  parent_folder = NULL,
  kind = NULL,
  owner = NULL
)
}
\arguments{
\item{conn}{Connection reference}

\item{pattern}{glob pattern for item name matching}

\item{parent_folder}{Numeric id of the containing folder}

\item{kind}{kind of item (optional)}

\item{owner}{owner of item (optional)}
}
\value{
Tibble of item properties
}
\description{
Retrieve BO object from an matching pattern
}
