% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-utils.R
\name{get_bo_item_id}
\alias{get_bo_item_id}
\title{Return Id from item properties or from numeric value}
\usage{
get_bo_item_id(item)
}
\arguments{
\item{item}{Item properties as tibble or item id as numeric value}
}
\value{
Numeric id of item
}
\description{
Return Id from item properties or from numeric value
}
