% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{get_bo_report_data}
\alias{get_bo_report_data}
\title{Get data from a BO document report tab}
\usage{
get_bo_report_data(conn, document, report, skip = 0)
}
\arguments{
\item{conn}{Connection reference obtained from open_bo_connection}

\item{document}{Document as numeric id or tibble of properties obtained from get_bo_item_from_name}

\item{report}{Name of report tab in document}

\item{skip}{Rows of data to skip before header}
}
\value{
Tibble of report data
}
\description{
Use get_bo_report_data to get data from a tab in a BO document. Be sure to save the report
after refreshing data in BO
}
\details{
Depending on the layout of the report tab you may need to skip rows in the output to get to the data
The data returned is the table as it was last saved in BO
}
