% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{get_bo_report_details}
\alias{get_bo_report_details}
\title{Get report details from document}
\usage{
get_bo_report_details(conn, document, report = "")
}
\arguments{
\item{conn}{Connection reference}

\item{document}{Document as numeric id or tibble of properties}

\item{report}{Name of report tab in document}
}
\value{
tibbe of report details
}
\description{
Get report details from document
}
