% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{get_bo_report_element_data}
\alias{get_bo_report_element_data}
\title{Title}
\usage{
get_bo_report_element_data(conn, document, report, element)
}
\arguments{
\item{conn}{Connection reference}

\item{document}{Document as numeric id or tibble of properties}

\item{report}{Name of report tab in document}

\item{element}{Name of element in report}
}
\value{
tibble with data from report element
}
\description{
Title
}
