% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{get_bo_report_inputs}
\alias{get_bo_report_inputs}
\title{Get inputs from a report table}
\usage{
get_bo_report_inputs(conn, document, inputs)
}
\arguments{
\item{conn}{Connection reference}

\item{document}{Document as numeric id or tibble of properties}

\item{inputs}{Name of report inputs as path "reports/report/element"}
}
\value{
Inputs as tibble
}
\description{
Get inputs from a report table
}
