% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spreadsheet.R
\name{get_bo_spreadsheet}
\alias{get_bo_spreadsheet}
\title{Get spreadsheet properties from BO}
\usage{
get_bo_spreadsheet(conn, filename, parent_folder)
}
\arguments{
\item{conn}{Connection reference}

\item{filename}{Name of the spreadsheet}

\item{parent_folder}{Numeric id of the folder}
}
\value{
List of properties
}
\description{
Get spreadsheet properties from BO
}
