% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{get_cached_token}
\alias{get_cached_token}
\title{Find a valid token matching server and username}
\usage{
get_cached_token(conn, server, username)
}
\arguments{
\item{conn}{Connection reference}

\item{server}{Server name and port}

\item{username}{Username}
}
\value{
TRUE if token found; FALSE otherwise
}
\description{
Find a valid token matching server and username
}
