% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{get_keyring_secret}
\alias{get_keyring_secret}
\title{Get a secret from the keyring}
\usage{
get_keyring_secret(name)
}
\arguments{
\item{name}{Secret name}
}
\value{
Secret value
}
\description{
Get a secret from the keyring
}
