% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/token-cache.R
\name{get_saved_items}
\alias{get_saved_items}
\title{Get items from table}
\usage{
get_saved_items(
  username = Sys.getenv("BO_USERNAME"),
  server = Sys.getenv("BO_SERVER"),
  table_name = get_token_table_name()
)
}
\arguments{
\item{table_name}{}
}
\description{
Get items from table
}
