% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/token-cache.R
\name{get_tokens}
\alias{get_tokens}
\title{Get tokens}
\usage{
get_tokens(username, server)
}
\arguments{
\item{username}{Name of the user}

\item{server}{Name of the server (host) with port}
}
\value{
Items matching keys
}
\description{
Get tokens
}
