% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{get_user_rights2}
\alias{get_user_rights2}
\title{Get user rights on the BO server}
\usage{
get_user_rights2(conn)
}
\arguments{
\item{conn}{connection reference}
}
\value{
Response content or NULL
}
\description{
Get user rights on the BO server
}
