% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{my_log_colorize}
\alias{my_log_colorize}
\title{Title}
\usage{
my_log_colorize(msg, level)
}
\arguments{
\item{msg}{The message to log}

\item{level}{The logging level}
}
\value{
nothing
}
\description{
Title
}
