% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{open_bo_connection}
\alias{open_bo_connection}
\title{Open a connection to a BO server}
\usage{
open_bo_connection(
  server = Sys.getenv("BO_SERVER"),
  username = Sys.getenv("BO_USERNAME"),
  password = NULL,
  conn = NULL
)
}
\arguments{
\item{server}{Server to connect to "server:port"}

\item{conn}{Connection reference to reuse (optional)}
}
\value{
Connection reference
}
\description{
Open a connection to a BO server
}
