% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{parse_bo_report_name}
\alias{parse_bo_report_name}
\title{To identify report for data, report name may have 'reports/' prefix}
\usage{
parse_bo_report_name(report)
}
\arguments{
\item{report}{Report name with possible 'reports/' prefix}
}
\value{
Report name without prefix
}
\description{
To identify report for data, report name may have 'reports/' prefix
}
