% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document.R
\name{POST_bo_document}
\alias{POST_bo_document}
\title{Upload a new webi document}
\usage{
POST_bo_document(conn, filepath, filename, folder_id)
}
\arguments{
\item{conn}{Connection reference}

\item{filepath}{Path to document (including filename)}

\item{filename}{name of document}

\item{folder_id}{Numeric id of parent folder for document}
}
\value{
Response content
}
\description{
Upload a new webi document
}
