% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spreadsheet.R
\name{post_bo_spreadsheet}
\alias{post_bo_spreadsheet}
\title{POST a new spreadsheet to BO folder}
\usage{
post_bo_spreadsheet(
  conn,
  filename,
  parent_folder,
  filepath = filename,
  format = "json"
)
}
\arguments{
\item{conn}{Connection reference}

\item{filename}{Name for spreasheet file}

\item{parent_folder}{Numeric id of the folder}

\item{filepath}{Path to spreadsheet file (including file name)}

\item{format}{HTTP body format to use}
}
\description{
POST a new spreadsheet to BO folder
}
