% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document.R
\name{refresh_bo_document}
\alias{refresh_bo_document}
\title{Title}
\usage{
refresh_bo_document(conn, document, dataSourceType = NULL)
}
\arguments{
\item{conn}{Connection reference}

\item{document}{Document as numeric id or tibble of properties}

\item{dataSourceType}{}
}
\description{
Title
}
