% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/token-cache.R
\name{remove_token}
\alias{remove_token}
\title{Remove tokens}
\usage{
remove_token(username, server)
}
\arguments{
\item{username}{Name of the user}

\item{server}{Name of the server (host) with port}
}
\description{
Remove tokens
}
