% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/token-cache.R
\name{save_token}
\alias{save_token}
\title{Get tokens}
\usage{
save_token(username, server, token)
}
\arguments{
\item{username}{Name of the user}

\item{server}{Name of the server (host) with port}

\item{token}{token to save}
}
\description{
Get tokens
}
