% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controls.R
\name{set_bo_control_selection}
\alias{set_bo_control_selection}
\title{Set the selected items for a control}
\usage{
set_bo_control_selection(
  conn,
  document,
  report = NULL,
  control,
  selections = NULL,
  all = FALSE
)
}
\arguments{
\item{conn}{Connection reference}

\item{document}{Document as numeric id or tibble of properties}

\item{report}{Identifier of report for report level controls, blank otherwise}

\item{control}{Identifier of the control}

\item{selections}{Selected items}

\item{all}{Select all items}
}
\value{
Response content
}
\description{
Set the selected items for a control
}
