% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controls.R
\name{set_bo_data_provider_specification}
\alias{set_bo_data_provider_specification}
\title{Set the date range provider specification}
\usage{
set_bo_data_provider_specification(conn, document, dataprovider, specification)
}
\arguments{
\item{conn}{Connection reference}

\item{document}{Document as numeric id or tibble of properties}

\item{dataprovider}{Name of data provider from document}

\item{specification}{XML specification of provider}
}
\description{
Set the date range provider specification
}
