% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document.R
\name{set_bo_document_control_selection}
\alias{set_bo_document_control_selection}
\title{Title}
\usage{
set_bo_document_control_selection(conn, document, controlName, selections, all)
}
\arguments{
\item{conn}{Connection reference}

\item{document}{Document as numeric id or tibble of properties}

\item{all}{}
}
\description{
Title
}
