% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document.R
\name{set_bo_document_data_source_date_range}
\alias{set_bo_document_data_source_date_range}
\title{Title}
\usage{
set_bo_document_data_source_date_range(
  conn,
  document,
  dataprovider,
  startDate,
  endDate
)
}
\arguments{
\item{conn}{Connection reference}

\item{document}{Document as numeric id or tibble of properties}

\item{dataprovider}{Name of data provider from document}

\item{startDate}{Start of date range}

\item{endDate}{End of date range}
}
\value{
Response content
}
\description{
Title
}
