% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{set_keyring_secret}
\alias{set_keyring_secret}
\title{Set a secret in the keyring}
\usage{
set_keyring_secret(name, value)
}
\arguments{
\item{name}{Secret name}

\item{value}{Secret value}
}
\description{
Set a secret in the keyring
}
