% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/token-cache.R
\name{set_user_password}
\alias{set_user_password}
\title{Set the password for a user in the connections database}
\usage{
set_user_password(
  username = Sys.getenv("BO_USERNAME"),
  server = Sys.getenv("BO_SERVER"),
  password
)
}
\arguments{
\item{username}{Name of the user}

\item{server}{Name of the server (host) with port}

\item{password}{Value for password}
}
\description{
Set the password for a user in the connections database
}
