% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{split_message}
\alias{split_message}
\title{Colorize a string by embedded tags starting with ';'}
\usage{
split_message(msg, level)
}
\arguments{
\item{msg}{The message to log}

\item{level}{The logging level}
}
\value{
Colorized string
}
\description{
Colorize a string by embedded tags starting with ';'
}
