% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spreadsheet.R
\name{upload_bo_agnostic_document}
\alias{upload_bo_agnostic_document}
\title{Upload a document of any type}
\usage{
upload_bo_agnostic_document(
  conn,
  filename,
  parent_folder,
  filepath = filename,
  type = NULL
)
}
\arguments{
\item{conn}{Connection reference}

\item{filename}{Name of file}

\item{parent_folder}{Parent folder to upload file to}

\item{type}{mime type (optional)}
}
\value{
result of upload
}
\description{
WIP
}
