% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spreadsheet.R
\name{upload_bo_dataframe}
\alias{upload_bo_dataframe}
\title{Upload a dataframe}
\usage{
upload_bo_dataframe(
  conn,
  df,
  filename,
  parent_folder,
  sheetname = "Sheet 1",
  type = NULL
)
}
\arguments{
\item{conn}{Connection reference}

\item{df}{Dataframe}

\item{filename}{Name of uploaded file}

\item{parent_folder}{Parent folder to upload file to}

\item{sheetname}{Name of the sheet}

\item{type}{of file (Optional, defauts to extension of filename)}
}
\value{
spreadsheat details
}
\description{
Upload a dataframe
}
