% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spreadsheet.R
\name{upload_df}
\alias{upload_df}
\title{Upload a dataframe}
\usage{
upload_df(conn, df, filename, parent_folder, sheetname = "Sheet 1")
}
\arguments{
\item{df}{Dataframe}

\item{filename}{Name of uploaded file}

\item{parent_folder}{Parent folder to upload file to}

\item{sheetname}{Name of the sheet}
}
\value{
NA
}
\description{
Upload a dataframe
}
