% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/borestapi-package.R
\docType{package}
\name{borestapi-package}
\alias{borestapi}
\alias{borestapi-package}
\title{borestapi: Interface to SAP Business Object REST API}
\description{
This package imports and exports data to SAP Business Objects. It fetches data from SAP Business Objects reports. It pushes excel files to SAP Business Objects local files.
}
\author{
\strong{Maintainer}: John Capehart \email{R@jcapehart.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
