% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/token-cache.R
\name{clear_user_password}
\alias{clear_user_password}
\title{Clear the password for a user in the connections database}
\usage{
clear_user_password(
  username = Sys.getenv("BO_USERNAME"),
  server = Sys.getenv("BO_SERVER")
)
}
\arguments{
\item{username}{Name of the user}

\item{server}{Name of the server (host) with port}
}
\description{
Clear the password for a user in the connections database
}
