% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{close_bo_connection}
\alias{close_bo_connection}
\title{Close connection to the BO server}
\usage{
close_bo_connection(conn)
}
\arguments{
\item{conn}{Connection reference to close}
}
\value{
Connection reference
}
\description{
Closes connection to the BO server.
}
\details{
As connection tokens are cached an reused to prevent flooding the BO
server with processes, it's not mandatory to call this.
}
