% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spreadsheet.R
\name{delete_bo_spreadsheet}
\alias{delete_bo_spreadsheet}
\title{Delete a spreadsheet from BO}
\usage{
delete_bo_spreadsheet(conn, filename, parent_folder)
}
\arguments{
\item{conn}{Connection reference}

\item{filename}{Name of the spreadsheet}

\item{parent_folder}{Numeric id of the folder}
}
\value{
Response content
}
\description{
Delete a spreadsheet from BO
}
