% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{get_bo_document_report}
\alias{get_bo_document_report}
\title{Get data from a BO document report tab}
\usage{
get_bo_document_report(conn, document, report, skip = 0)
}
\arguments{
\item{conn}{Connection reference obtained from open_bo_connection}

\item{document}{Document as numeric id or tibble of properties obtained from get_bo_item_from_name}

\item{report}{Name of report tab in document}

\item{skip}{Rows of data to skip before header in the case where the report tab in business objects
has text lines before the data}
}
\value{
Tibble of report data
}
\description{
Get data from a BO document report tab
}
