% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spreadsheet.R
\name{GET_bo_spreadsheet}
\alias{GET_bo_spreadsheet}
\title{Get spreadsheet properties from BO}
\usage{
GET_bo_spreadsheet(conn, filename, folder_id)
}
\arguments{
\item{conn}{Connection reference}

\item{filename}{Name of the spreadsheet}

\item{folder_id}{Numeric id of the folder}
}
\value{
List of properties
}
\description{
Get spreadsheet properties from BO
}
