% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/token-cache.R
\name{get_user_password}
\alias{get_user_password}
\title{Title}
\usage{
get_user_password(
  username = Sys.getenv("BO_USERNAME"),
  server = Sys.getenv("BO_SERVER")
)
}
\arguments{
\item{username}{Name of the user}

\item{server}{Name of the server (host) with port}
}
\value{
Password
}
\description{
Title
}
