% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{open_bo_connection}
\alias{open_bo_connection}
\title{Open a connection to a BO server.}
\usage{
open_bo_connection(
  server = Sys.getenv("BO_SERVER"),
  username = Sys.getenv("BO_USERNAME"),
  password = get_user_password(username, server),
  save_password = TRUE
)
}
\arguments{
\item{server}{Server to connect to "server:port" (optional)
Defaults 'BO_SERVER' environment variable}

\item{username}{(optional) Defaults to 'BO_USERNAME' environment variable}

\item{password}{(optional) Defaults to password previously saved for server username combination}

\item{save_password}{Save the password encrypted in the database.
Saving password is required for automatic reconnection on token expiration}
}
\value{
Connection reference
}
\description{
Defaults can be set in environment variables
}
\examples{
# open a connection to the server in environment variable "BO_SERVER" using user in "BO_USERNAME".
# and password in the database
conn <- open_bo_connection()
}
