% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{open_bo_connection}
\alias{open_bo_connection}
\title{Open a connection to a BO server. You can keep parameters in environment variables}
\usage{
open_bo_connection(
  server = Sys.getenv("BO_SERVER"),
  username = Sys.getenv("BO_USERNAME"),
  password = NULL,
  conn = NULL
)
}
\arguments{
\item{server}{Server to connect to "server:port" (optional). Defaults BO_SERVER environment variable}

\item{username}{(optional). Defaults to "BO_USERNAME environment variable}

\item{password}{(optional). Use getPass() to enter value}

\item{conn}{Connection reference to reuse (optional)}
}
\value{
Connection reference
}
\description{
Open a connection to a BO server. You can keep parameters in environment variables
}
\examples{
# open a connection to the server in environment variable "BO_SERVER" using user in "BO_USERNAME". and password in a
# keyring at "BO_KEYRING_FILE_PATH"
open_bo_connection()
# open a connection to a server using username and password
open_bo_connection(Sys.getenv("BO_SERVER"), username = 'john.capehart', password=getPass())
}
