% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document.R
\name{post_bo_document}
\alias{post_bo_document}
\title{Upload a new webi document}
\usage{
post_bo_document(conn, filepath, filename, parent_folder)
}
\arguments{
\item{conn}{Connection reference}

\item{filepath}{Path to document (including filename)}

\item{filename}{name of document}

\item{parent_folder}{Numeric id of parent folder for document}
}
\value{
Response content
}
\description{
Upload a new webi document
}
