% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spreadsheet.R
\name{POST_bo_spreadsheet_raw}
\alias{POST_bo_spreadsheet_raw}
\title{POST a new spreadsheet to BO folder}
\usage{
POST_bo_spreadsheet_raw(conn, filepath, filename, folder_id, format = "json")
}
\arguments{
\item{conn}{Connection reference}

\item{filepath}{Path to spreadsheet file (including name)}

\item{filename}{Name for spreasheet file}

\item{folder_id}{Numeric id of the folder}

\item{format}{HTTP body format to use}
}
\value{
Response content
}
\description{
POST a new spreadsheet to BO folder
}
