% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document.R
\name{put_bo_document}
\alias{put_bo_document}
\title{Upload an existing BO Webi document}
\usage{
put_bo_document(conn, filepath, filename, document_id)
}
\arguments{
\item{conn}{Connection reference}

\item{filepath}{Path to document (including filename)}

\item{filename}{name of document}

\item{document_id}{Numeric id of Webi document}
}
\value{
Response content
}
\description{
Upload an existing BO Webi document
}
