% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spreadsheet.R
\name{PUT_bo_spreadsheet}
\alias{PUT_bo_spreadsheet}
\title{PUT an existing spreadsheet to BO folder}
\usage{
PUT_bo_spreadsheet(conn, filepath, filename, sheetId)
}
\arguments{
\item{conn}{Connection reference}

\item{filepath}{Path to spreadsheet file (including file name)}

\item{filename}{Name for spreasheet file}

\item{folder_id}{Numeric id of the folder}
}
\description{
PUT an existing spreadsheet to BO folder
}
