% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spreadsheet.R
\name{put_bo_spreadsheet}
\alias{put_bo_spreadsheet}
\title{PUT an existing spreadsheet to BO folder}
\usage{
put_bo_spreadsheet(conn, filename, sheet_id, filepath = filename)
}
\arguments{
\item{conn}{Connection reference}

\item{filename}{Name for spreasheet file}

\item{sheet_id}{Numeric id of spreadsheet document}

\item{filepath}{Path to spreadsheet file (including file name)}
}
\description{
PUT an existing spreadsheet to BO folder
}
