% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document.R
\name{refresh_bo_data_provider}
\alias{refresh_bo_data_provider}
\title{Title}
\usage{
refresh_bo_data_provider(conn, document, data_provider = NULL)
}
\arguments{
\item{conn}{Connection reference}

\item{document}{Document as numeric id or tibble of properties}

\item{data_provider}{Identifier of the data provider}
}
\value{
Response content
}
\description{
Title
}
