% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/token-cache.R
\name{set_user_password}
\alias{set_user_password}
\title{Title}
\usage{
set_user_password(
  username = Sys.getenv("BO_USERNAME"),
  server = Sys.getenv("BO_SERVER"),
  password
)
}
\arguments{
\item{password}{}
}
\description{
Title
}
