% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spreadsheet.R
\name{upload_bo_spreadsheet}
\alias{upload_bo_spreadsheet}
\title{Title}
\usage{
upload_bo_spreadsheet(conn, filepath, filename, folder_id)
}
\arguments{
\item{conn}{Connection reference}

\item{filepath}{Path to spreadsheet file (including file name)}

\item{filename}{Name of the spreadsheet}

\item{folder_id}{Numeric id of the folder}
}
\value{
List of properties for spreadsheet
}
\description{
Title
}
