% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spreadsheet.R
\name{upload_bo_spreadsheet}
\alias{upload_bo_spreadsheet}
\title{Upload a spreadsheet}
\usage{
upload_bo_spreadsheet(
  conn,
  filename,
  parent_folder,
  filepath = filename,
  type = NULL
)
}
\arguments{
\item{conn}{Connection reference}

\item{filename}{Name of the spreadsheet}

\item{parent_folder}{Numeric id of the folder}

\item{filepath}{Path to spreadsheet file (including file name). Defaults to filename}
}
\value{
List of properties for spreadsheet
}
\description{
Upload a spreadsheet
}
